[CPDLP-1691] Improve new api v3 partnership endpoint namespacing
