+++ app/controllers/concerns/api_orderable.rb
@@ -3,17 +3,9 @@
 module ApiOrderable
   extend ActiveSupport::Concern
 
-  module ClassMethods
-    attr_reader :model_klass
+private
 
-  private
-
-    def orderable(model_klass: nil)
-      @model_klass = model_klass
-    end
-  end
-
-  def sort_params(params)
+  def sort_params(params, model: controller_name.classify.constantize)
     sort = {}
     if params[:sort]
       sort_order = { "+" => :asc, "-" => :desc }
@@ -22,7 +14,6 @@ module ApiOrderable
       sorted_params.each do |attr|
         sort_sign = attr =~ /\A[+-]/ ? attr.slice!(0) : "+"
 
-        model = self.class.model_klass.presence || controller_name.classify.constantize
         if model.attribute_names.include?(attr)
           sort[attr] = sort_order[sort_sign]
         end
