+++ config/routes.rb
@@ -130,7 +130,7 @@ Rails.application.routes.draw do
     namespace :v3, constraints: ->(_request) { FeatureFlag.active?(:api_v3) } do
       resources :statements, only: %i[index show], controller: "finance/statements"
       resources :delivery_partners, only: %i[index show], path: "delivery-partners"
-      resources :partnerships, path: "partnerships/ecf", only: %i[index], controller: "ecf_partnerships"
+      resources :partnerships, path: "partnerships/ecf", only: %i[index], controller: "ecf/partnerships"
     end
   end
