+++ spec/docs/v3/ecf_partnerships_spec.rb
@@ -21,6 +21,18 @@ RSpec.describe "API", type: :request, swagger_doc: "v3/api_spec.json" do
                 description: "Refine partnerships to return.",
                 example: "filter[cohort]=2021,2022"
 
+      parameter name: :sort,
+                in: :query,
+                schema: {
+                  "$ref": "#/components/schemas/PartnershipsSort",
+                },
+                type: :array,
+                style: :form,
+                explode: false,
+                required: false,
+                description: "Sort partnerships being returned.",
+                example: "sort=-updated_at"
+
       response "200", "A list of ECF partnerships" do
         schema({ "$ref": "#/components/schemas/MultipleECFPartnershipsResponse" })
