+++ swagger/v3/component_schemas/PartnershipsSort.yml
@@ -1,0 +1,9 @@
+description: "Sort partnerships being returned"
+type: array
+items:
+  type: string
+  enum:
+    - created_at
+    - -created_at
+    - updated_at
+    - -updated_at
